requirements.txt
appdirs==1.4.4
asgiref==3.4.1
azure-core==1.20.1
azure-storage-blob==12.9.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.7
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
distlib==0.3.3
dj-database-url==0.5.0
Django==3.1.8
django-cors-headers==3.7.0
django-environ==0.4.5
django-rest-swagger==2.2.0
django-rq==2.4.1
django-storages==1.12.3
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filelock==3.3.2
flake8==3.8.4
greenlet==1.1.2
gunicorn==20.1.0
identify==2.3.5
idna==3.3
inflection==0.5.1
isodate==0.6.0
itypes==1.2.0
Jinja2==3.0.3
jsonfield==3.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
msrest==0.6.21
nodeenv==1.6.0
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.2
platformdirs==2.4.0
pre-commit==2.12.1
pycodestyle==2.6.0
psycopg2-binary==2.9.2
pycparser==2.21
pydocstyle==6.0.0
pyodbc==4.0.32
pyflakes==2.2.0
Pygments==2.8.0
PyJWT==2.3.0
pymssql==2.2.2
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.3
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rq==1.10.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
snowballstemmer==2.1.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
